account/authentication.py
# account/authentication.py
from rest_framework_simplejwt.authentication import JWTAuthentication

class CookieJWTAuthentication(JWTAuthentication):
    def authenticate(self, request):
        token = request.COOKIES.get('access_token')
        if not token:
            return None
        request.META['HTTP_AUTHORIZATION'] = f'JWT {token}'
        return super().authenticate(request)
