requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.11.1
boto3==1.21.13
botocore==1.24.13
certifi==2021.10.8
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.1.3
constantly==15.1.0
coverage==6.2
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
distlib==0.4.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==22.1
django-jsonfield==1.4.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.20.0
Flask==2.2.2

hiredis==3.2.1
httpie==2.6.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.12.0
incremental==21.3.0
iso8601==0.1.12
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
MouseInfo==0.1.3
msgpack==1.0.2
pillow==11.3.0
platformdirs==4.5.0
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==2.3.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.1
pywhatkit==5.4
qrcode==8.2
redis==3.5.3
requests==2.26.0
requests-toolbelt==0.9.1
s3transfer==0.5.2
service-identity==21.1.0
setuptools==80.9.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
twilio==7.9.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.35.3
Werkzeug==2.2.2
wheel==0.45.1
wikipedia==1.4.0
zipp==3.8.1
zope.interface==5.4.0
